// src/app/models/patient.model.ts
export class Patient {
    id: string;
    name: string;
    surname: string;
    email: string;
    tel: string;
    dob: string;
    photo: string;
    address: string;
    postal_code: string;
    country: string;
    city: string;
    username: string;
    gender: string;
    password: string;
    certify: string;
    creation_date: string;
  }